refactor: mypage-약관페이지 ssr 코드분리
